fix: tile h style and byline
