+++ packages/article-byline/article-byline-opinion.showcase.js
@@ -5,12 +5,6 @@ import { ArticleBylineOpinion } from "./src/article-byline";
 
 const authorsAST = require("./fixtures/authors.json");
 
-const styles = {
-  fontSize: 12,
-  letterSpacing: 0.6,
-  lineHeight: 12
-};
-
 const preventDefaultedAction = decorateAction =>
   decorateAction([
     ([e, ...args]) => {
@@ -39,19 +33,6 @@ export default {
         </ComponentWrapper>
       ),
       name: "Byline with a single author",
-      type: "story"
-    },
-    {
-      component: (_, { decorateAction }) => (
-        <ComponentWrapper>
-          <ArticleBylineOpinion
-            ast={authorsAST.authorInTheBeginning}
-            {...getProps(decorateAction)}
-            opinionStyle={styles}
-          />
-        </ComponentWrapper>
-      ),
-      name: "Byline with a single author and style overriden",
       type: "story"
     }
   ],
