+++ packages/article-byline/src/article-byline.js
@@ -4,9 +4,18 @@ import renderByline from "./render-byline";
 import { propTypes, defaultProps } from "./article-byline-prop-types";
 import styles from "./styles";
 
-const renderAuthorComponent = (children, key, attributes, { className }) => ({
+const renderAuthorComponent = (
+  children,
+  key,
+  attributes,
+  { className, opinionStyle }
+) => ({
   element: (
-    <Text className={className} key={key} style={styles.nonLinkText}>
+    <Text
+      className={className}
+      key={key}
+      style={[styles.nonLinkText, opinionStyle]}
+    >
       {children}
     </Text>
   )
